VideoPress: Move use-poster-edit back to admin
